Move enrich views to sdk
