+++ packages/server/src/sdk/app/tables/index.ts
@@ -5,9 +5,18 @@ import {
   isExternalTable,
   isSQL,
 } from "../../../integrations/utils"
-import { Table, Database } from "@budibase/types"
+import {
+  Table,
+  Database,
+  TableResponse,
+  TableViewsResponse,
+  TableSchema,
+  UIFieldMetadata,
+} from "@budibase/types"
 import datasources from "../datasources"
 import { populateExternalTableSchemas, isEditableColumn } from "./validation"
+import sdk from "../../../sdk"
+import _ from "lodash"
 
 async function getAllInternalTables(db?: Database): Promise<Table[]> {
   if (!db) {
@@ -55,6 +64,47 @@ async function getTable(tableId: any): Promise<Table> {
   }
 }
 
+function enrichSchema(
+  tableSchema: TableSchema,
+  viewOverrides: Record<string, UIFieldMetadata>
+) {
+  const result: TableSchema = {}
+  for (const [columnName, columnUIMetadata] of Object.entries(viewOverrides)) {
+    if (!columnUIMetadata.visible) {
+      continue
+    }
+
+    if (!tableSchema[columnName]) {
+      continue
+    }
+
+    result[columnName] = _.merge(tableSchema[columnName], columnUIMetadata)
+  }
+  return result
+}
+
+function enrichViewSchemas(table: Table): TableResponse {
+  const result: TableResponse = {
+    ...table,
+    views: Object.values(table.views ?? []).reduce((p, v) => {
+      if (!sdk.views.isV2(v)) {
+        p[v.name] = v
+      } else {
+        p[v.name] = {
+          ...v,
+          schema:
+            !v?.columns || !Object.entries(v?.columns).length
+              ? table.schema
+              : enrichSchema(table.schema, v.columns),
+        }
+      }
+      return p
+    }, {} as TableViewsResponse),
+  }
+
+  return result
+}
+
 export default {
   getAllInternalTables,
   getAllExternalTables,
@@ -62,4 +112,5 @@ export default {
   getTable,
   populateExternalTableSchemas,
   isEditableColumn,
+  enrichViewSchemas,
 }
